Rename ManageRemotes -> ManagePeers
