+++ ui/src/path.ts
@@ -38,7 +38,7 @@ export const projectCommits = (id: string): string => `/projects/${id}/commits`;
 export const designSystemGuide = (): string => "/design-system-guide";
 
 // modal routes
-export const manageRemotes = (): string => "/manage-remotes";
+export const managePeers = (): string => "/manage-peers";
 export const newProject = (): string => "/new-project";
 export const search = (): string => "/search";
 export const shortcuts = (): string => "/shortcuts";
